requirements.txt
annotated-types==0.7.0
babel==2.17.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
courlan==1.3.2
dateparser==1.2.1
feedparser==6.0.11
Flask==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
gunicorn==23.0.0
htmldate==1.9.3
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
json5==0.12.0
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
sgmllib3k==1.0.0
six==1.17.0
tld==0.13.1
tqdm==4.67.1
trafilatura==2.0.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
Werkzeug==3.1.3
